requirements.txt
# Add your project dependencies here
streamlit
langchain
langchain-community
langchain-groq
sentence-transformers
faiss-cpu
python-dotenv

